Create Book allocation problem
Give or an arrays of integer number n . where a[i]  represent the number of  pages in  the book .  There are 'm' number of student and task is  to allocate all the book  to their student .  Allocate book in such a way. 
1). each student get at least one book.
2). each book should be allocate to student. 
3).Book allocation should be in contiguous manner.

You  have to allocate  the book  to m 'student such that  the maximum  number of pages assigned  to a student minimum. 

example :- let consider n=4 (number of book ) and 2 students. 
arr=[10,20,30,40];
than answer is 60;