Create getter and satter method
This method is uses to access private data mamber of the class by object outside the object. 

A getter is a method that retrieves (gets) the value of a private class member.

A setter is a method that updates (sets) the value of a private class member.

They are used because in Object-Oriented Programming (OOP), class data members are usually declared private (hidden), and you control access to them through public functions.